Wifi tracking working!
The worst of it is it does have to be connected to a network to work the
way I want.  Attempts to plow through the source code came up nada.  I'm
leaving it as is (it works with the envisioned goals) and moving on to
the actual webpage portion of it.  Time to merge with the master