First attempts at the phone sync
I don't have the process quite down yet.  The wifi works, the phone can
connect, log in, load the page, then ask the device to send a message.
However, connection to a server is troublesome and it breaks the phone
connection when it does it.

I may abandon this feature in favour of focusing on the RFID.  I just
need the transistors to come in so I can wire everything together (I
don't have a spare RFID reader to use as a prototype on the breadboard)