Pin layout changes and Quest System update
It's still not finished but I need to move the code up to have access to
it.

The pins are properly set though not properly documented. Still a little
up in the air.  It looks like I'll have to do the LCD CS as the RFID
keeps causing abberations on one of the screens.